SolanaYieldOrchestrator/requirements.txt
flask==3.0.0
flask-cors==4.0.0
solana==0.34.3
solders==0.21.0
driftpy==0.8.82
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0
anchorpy==0.20.1
redis==5.0.1
aiohttp
hyperliquid-python-sdk
solana
solders
websockets
solana-agent-kit
